requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scikit-image>=0.18.0
tensorflow>=2.8.0
matplotlib>=3.4.0
seaborn>=0.11.0
Pillow>=8.3.0
joblib>=1.0.0
kaggle>=1.5.12
xgboost>=1.5.0